Test fix

+++ include/battle_ai_script_commands.h
@@ -7,9 +7,9 @@
 #define AI_CHOICE_WATCH 5
 #define AI_CHOICE_SWITCH 7
 
-s32 AI_CalcDamage(u16 move, u16 battlerAtk, u16 battlerDef);
-s32 AI_CalcPartyMonDamage(u16 move, u16 battlerAtk, u16 battlerDef, struct Pokemon *mon);
-u16 AI_GetTypeEffectiveness(u16 move, u16 battlerAtk, u16 battlerDef);
+s32 AI_CalcDamage(u16 move, u8 battlerAtk, u8 battlerDef);
+s32 AI_CalcPartyMonDamage(u16 move, u8 battlerAtk, u8 battlerDef, struct Pokemon *mon);
+u16 AI_GetTypeEffectiveness(u16 move, u8 battlerAtk, u8 battlerDef);
 void BattleAI_SetupItems(void);
 void BattleAI_SetupFlags(void);
 void BattleAI_SetupAIData(u8 defaultScoreMoves);
